Add the ability to instrument a graphql request
A Chain of Responsibility pattern is employed to allow
callers adding custom instrumentation targeting graphql
request and response objects.
A couple good use cases for this are request logging,
caching, and certain kinds of auth (e.g. bearer token renewal).

In order to proceed of this kind of support a small refactor
was required. So, there are somee changes that impact directly
the previous structure of this module.